style: format with forge fmt

+++ src/fundme.sol
@@ -17,7 +17,7 @@ contract FundMe {
     address[] private s_funders;
 
     // Could we make this constant?  /* hint: no! We should make it immutable! */
-    address private immutable  i_owner;
+    address private immutable i_owner;
     uint256 public constant MINIMUM_USD = 5 * 10 ** 18;
 
     AggregatorV3Interface private s_pricefeed;
@@ -45,11 +45,11 @@ contract FundMe {
     }
 
     function cheapwithdraw() public onlyOwner {
-       uint funderslength = s_funders.length;//only read s_funders.length once 
-         for (uint256 funderIndex = 0; funderIndex < funderslength; funderIndex++) {
-              address funder = s_funders[funderIndex];
-              s_addressToAmountFunded[funder] = 0;
-         }
+        uint256 funderslength = s_funders.length; //only read s_funders.length once
+        for (uint256 funderIndex = 0; funderIndex < funderslength; funderIndex++) {
+            address funder = s_funders[funderIndex];
+            s_addressToAmountFunded[funder] = 0;
+        }
         s_funders = new address[](0);
         (bool callSuccess,) = payable(msg.sender).call{value: address(this).balance}("");
         require(callSuccess, "Call failed");
